+++ kube/src/lib.rs
@@ -97,7 +97,7 @@ mod oauth2;
 
 #[cfg(feature = "derive")] pub use kube_derive::CustomResource;
 
-pub use api::{Api, Resource};
+pub use api::{Api, Resource, DynamicResource};
 #[doc(inline)] pub use client::Client;
 #[doc(inline)] pub use config::Config;
 #[doc(inline)] pub use error::Error;